Prep gem with dependencies and test suite setup
This preps the gem by adding it's dependencies and adding a
db schema context for the test suite. The intent with the context
is to test with ActiveRecord and a SQLite in-memory database.
I added ActiveRecord as a dependency because that is the intent
of this gem, to wrap ActiveRecord. This narrows the support scope
for this gem and allows us to specifically support ActiveRecord
and not a myriad of similar libraries.