Add fake record assertion helpers
This adds helpers for asserting that actions occurred on a fake
record. Specifically this adds a helper for determining if a
field was saved (or not saved) and if a fake was destroyed (or
not). This allows using a fake record with a model and testing
that the model was saved or destroyed.