Allow writing has many associations on a `MR::Model`
This adds writer's for has many associations defined on a
`MR::Model`. This was already done for belongs to associations
but I didn't have a use-case and wanted to limit the methods
available on a `MR::Model`. I typically don't like writing
has many associations (ActiveRecord kinda magically takes care
of it) but it fits well with the practice of not
interacting with id's directly. The primary use for this is in
tests when you want to setup a model and it's associated models
using fakes.