modify test helpers assertion methods to use Assert's `with_backtrace`
This helper was added in Assert v2.2.0 (redding/assert#133). Using
this helpers allows the results generated in the test helpers to have
the caller's backtrace, which is more intuitive for the user.

Notice, too, that `with_backtrace` is used in the context methods
used to test that `with_backtrace` is being used in the test helpers -
SO META!