Add first pass `MR::ReadModel`
This adds a first pass at defining a `MR::ReadModel`. The goal
is to be able to use a read model instead of an actual MR model,
and have it provide the interface for reading values from a
record.

The primary use-case for this is when doing an AR scope with
custom SELECT columns. Currently, MR doesn't provide a way to
get these, and they don't fit well with the MR model.
`MR::ReadModel` handles this. It can be used with `MR::Query`
and will take a record, similar to `MR::Model` and expose it's
attributes (even custom selects) as needed.

Closes #23
Closes #41