Organize factories into sub-files, for easier readability
The record and model factory classes were getting lost in the
main factory file, so I wanted to split them out similar to how
stacks are divided up. This makes it easier to read their code
and find the classes. I also renamed them to have the `Factory`
suffix, as there are a lot of "models" and "records" in MR and I
wanted it to be easy to identify what these were.