Allow calling `transaction` on a model class
Similar to how you can call `transaction` on a record class, this
allows calling it on the model class. This is a pass through to
the record class's transaction method. This is for convenience
when you are wanting to update a set of models and don't have a
specific instance for initiating the transaction. Functionally,
this works the exact same as calling a transaction method on an
ActiveRecord class.