Add a benchmarking script
This adds a benchmarking script for MR models and read models.
This may be expanded to more parts of MR in the future. This is
intended as a simple way to relatively detect changes in speed
and memory usage of MR. This will be used shortly to detect if
changes to MR are more or less efficient than they previously
were. Each piece is benchmarked in as much isolation as possible.