Add new interface for read model querying, only basics for now
This adds a new mixin `MR::ReadModel::Querying`. This mixin will
be used to update read models to provide a fully-featured API
for querying data from the DB. For now this only implements the
basic interface and tests it. This only includes the DSL methods
`from`, `select` and `joins` and the `query` method. The `from`
method takes a record class to build the relation from and the
`select` and `joins` methods map to ActiveRecord's querying.
This doesn't make any changes to the `MR::ReadModel` mixin yet,
this is just setup for changing it.

Related to #88