Use the test DB setup in the bench scripts, split to many
This updates the bench scripts to use the test DB setup. It also
changes them to be multiple scripts and to use the `whysoslow`
gem. This reduces the code-debt of the bench script and makes it
more friendly to work with.

**Note** This resets the benchmarks for models and read models.
No code was changed, only the benchmarks were changed.

Closes #98
Closes #121
Closes #122
Closes #123