Remove model associations validation
This removes the validation that was added for model
assocaitions. This was a nice help-the-developer-out feature, but
ultimately isn't very flexible with model mixins that add
associations. It requires the specific order of setting the
record class before adding any associations, which isn't bad, but
can be annoying when working with mixins. Ultimately, the benefit
of MR validating that an association exists on it's record class
(which a test suite should catch) is not worth the degradation in
flexibility.