Use Ardb for setting up ActiveRecord with a test DB
This switches to using Ardb to setup ActiveRecord with a test
DB connection. Ardb will always handle doing this, no matter how
ActiveRecord changes, so it's better to move to using it. It's
also simpler and provides a cleaner solution.

This also cleans up the duplicate ActiveRecord `Relation` fake and
spy in the test suite. This is setup to pull the spy out and move
it into Ardb. This is the proper place for ActiveRecord spys and
fakes like this.

Related to #98