Remove shared support models for unit tests
This removes 2 shared support models that multiple unit tests
were using. For me, this has become an anti-pattern. These
"support" files need to be different things for each unit test,
depending on what it needs. They are intended to be specific for
the unit tests, so they shouldn't live in a common place. When
they are commonized, it encourages coupling of tests, which makes
them harder to maintain and the support files has to make all of
it's use-cases work together. All of this is more overhead than
the commonization is worth.

This also adds an active support require to the fake record
associations. When the tests for the associations are run in
isolation, they fail because `constantize` from active support
wasn't available.