Updates to support rubinius 2.2.1 and MRI 2.0.0
These are some minor updates to make sure that MR is compatible
with rubinius 2.2.1 and MRI 2.0.0. This doesn't change any of
the functionality of MR and just makes sure the test suite passes
and the bench scripts run. This consisted of the following
changes:

* Add `rubysl` to the Gemfile. This is how you setup using
  ruby's standard library for rubinius.
* The gem's folder needed to be added as a load path for the
  bench scripts.
* In fake record associations tests, updated a spot that was
  creating an association. The second arg is expected to be
  a `Hash`, not a `String`. Somehow, this didn't throw an
  exception in REE 1.8.7. I believe it was a mix of `String`
  responding to `[]` and it somehow implicitly converting
  a `Symbol` into an `Integer`. Crazy.
* In read model tests, updated a test that was depending on
  MRI's `Array#to_s`. In rubinius, `Array#to_s` is more like
  `inspect`. Manually joined the array to build the desired
  string.

This includes a bash script called `test_all` that will run all
of the supported versions using `rb`. This also sets up MR to run
on Travis CI. This is an experiment to see how Travis CI works for
continous integration.