Add `MR::FakeRecord::Persistence` mixin
This adds a mixin for fake records that contains all the
"persistence" related logic for fake records. This is to better
organize the behavior of fake records and make it easier to test.
This is an internal change and doesn't make any changes to the
fake record's behavior yet.

This also requires thread in the factory. This should've been
done since it uses `Mutex` for locking.