Add `assert_association_saved` test helpers
This adds `assert_association_saved` test helpers which can be
used to test that an association was saved. These use the field
saved assertions to check that the association's foreign key was
saved. For polymorphic associations, it will test that both the
foreign key and foreign type were saved. All of this was already
available using `assert_field_saved`, this is just more convenient
and expressive.

Closes #89